backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
google-generativeai>=0.5.0
requests
pydantic
certifi
